refactor aggregation functions code structure.
fix format.

fix minimal compile

fix format

fix conflicts.
