+++ velox/exec/tests/CMakeLists.txt
@@ -86,7 +86,6 @@ add_test(
 target_link_libraries(
   velox_exec_test
   velox_aggregates
-  velox_aggregates_test_lib
   velox_dwio_common
   velox_dwio_common_exception
   velox_dwio_common_test_utils
