+++ velox/functions/lib/aggregates/CMakeLists.txt
@@ -12,20 +12,11 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
-add_executable(
-  velox_functions_spark_aggregates_test
-  BitwiseXorAggregationTest.cpp FirstAggregateTest.cpp LastAggregateTest.cpp
-  Main.cpp)
+add_library(velox_functions_aggregates SingleValueAccumulator.cpp)
 
-add_test(velox_functions_spark_aggregates_test
-         velox_functions_spark_aggregates_test)
+target_link_libraries(velox_functions_aggregates velox_exec
+                      velox_presto_serializer ${FOLLY_WITH_DEPENDENCIES})
 
-target_link_libraries(
-  velox_functions_spark_aggregates_test
-  velox_exec_test_lib
-  velox_vector_test_lib
-  velox_functions_aggregates_test_lib
-  velox_functions_spark_aggregates
-  gflags::gflags
-  gtest
-  gtest_main)
+if(${VELOX_BUILD_TESTING})
+  add_subdirectory(tests)
+endif()
