+++ velox/functions/lib/aggregates/SimpleNumericAggregate.h
@@ -21,7 +21,7 @@
 #include "velox/vector/FlatVector.h"
 #include "velox/vector/LazyVector.h"
 
-namespace facebook::velox::aggregate {
+namespace facebook::velox::functions::aggregate {
 
 template <typename TInput, typename TAccumulator, typename TResult>
 class SimpleNumericAggregate : public exec::Aggregate {
@@ -96,7 +96,7 @@ class SimpleNumericAggregate : public exec::Aggregate {
     DecodedVector decoded(*arg, rows, !mayPushdown);
     auto encoding = decoded.base()->encoding();
     if (encoding == VectorEncoding::Simple::LAZY) {
-      SimpleCallableHook<TValue, TData, UpdateSingleValue> hook(
+      velox::aggregate::SimpleCallableHook<TValue, TData, UpdateSingleValue> hook(
           exec::Aggregate::offset_,
           exec::Aggregate::nullByte_,
           exec::Aggregate::nullMask_,
@@ -240,4 +240,4 @@ class SimpleNumericAggregate : public exec::Aggregate {
   }
 };
 
-} // namespace facebook::velox::aggregate
+} // namespace facebook::velox::functions::aggregate
