+++ velox/functions/prestosql/aggregates/tests/ApproxDistinctTest.cpp
@@ -15,10 +15,11 @@
  */
 #include "velox/common/hyperloglog/HllUtils.h"
 #include "velox/exec/tests/utils/PlanBuilder.h"
-#include "velox/functions/prestosql/aggregates/tests/AggregationTestBase.h"
+#include "velox/functions/lib/aggregates/tests/AggregationTestBase.h"
 
 using namespace facebook::velox::exec;
 using namespace facebook::velox::exec::test;
+using namespace facebook::velox::functions::aggregate::test;
 
 namespace facebook::velox::aggregate::test {
 namespace {
