+++ velox/functions/prestosql/aggregates/tests/BoolAndOrTest.cpp
@@ -15,8 +15,9 @@
  */
 #include "velox/common/base/tests/GTestUtils.h"
 #include "velox/exec/tests/utils/PlanBuilder.h"
-#include "velox/functions/prestosql/aggregates/tests/AggregationTestBase.h"
+#include "velox/functions/lib/aggregates/tests/AggregationTestBase.h"
 
+using namespace facebook::velox::functions::aggregate::test;
 using facebook::velox::exec::test::PlanBuilder;
 
 namespace facebook::velox::aggregate::test {
