+++ velox/functions/prestosql/aggregates/tests/CMakeLists.txt
@@ -12,10 +12,6 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
-add_library(velox_aggregates_test_lib AggregationTestBase.cpp)
-
-target_link_libraries(velox_aggregates_test_lib velox_exec_test_lib)
-
 add_executable(
   velox_aggregates_test
   AggregationFunctionRegTest.cpp
@@ -52,12 +48,13 @@ add_test(
 target_link_libraries(
   velox_aggregates_test
   velox_aggregates
-  velox_aggregates_test_lib
   velox_core
   velox_dwio_common_test_utils
   velox_exec
   velox_exec_test_lib
   velox_file
+  velox_functions_aggregates
+  velox_functions_aggregates_test_lib
   velox_functions_test_lib
   velox_functions_prestosql
   velox_functions_lib
