+++ velox/functions/prestosql/aggregates/tests/ChecksumAggregateTest.cpp
@@ -16,10 +16,11 @@
 #include <boost/algorithm/string/join.hpp>
 #include "velox/common/encode/Base64.h"
 #include "velox/exec/tests/utils/PlanBuilder.h"
-#include "velox/functions/prestosql/aggregates/tests/AggregationTestBase.h"
+#include "velox/functions/lib/aggregates/tests/AggregationTestBase.h"
 #include "velox/functions/prestosql/types/TimestampWithTimeZoneType.h"
 
 using namespace facebook::velox::exec::test;
+using namespace facebook::velox::functions::aggregate::test;
 
 namespace facebook::velox::aggregate::test {
