+++ velox/functions/sparksql/aggregates/BitwiseXorAggregate.h
@@ -18,8 +18,8 @@
 
 #include <string>
 
-namespace facebook::velox::functions::sparksql::aggregates {
+namespace facebook::velox::functions::aggregate::sparksql {
 
 bool registerBitwiseXorAggregate(const std::string& name);
 
-} // namespace facebook::velox::functions::sparksql::aggregates
+} // namespace facebook::velox::functions::aggregate::sparksql
