+++ velox/functions/sparksql/aggregates/Register.cpp
@@ -18,12 +18,12 @@
 
 #include "velox/functions/sparksql/aggregates/BitwiseXorAggregate.h"
 
-namespace facebook::velox::functions::sparksql::aggregates {
+namespace facebook::velox::functions::aggregate::sparksql {
 
 extern void registerFirstLastAggregates(const std::string& prefix);
 
 void registerAggregateFunctions(const std::string& prefix) {
   registerFirstLastAggregates(prefix);
-  aggregates::registerBitwiseXorAggregate(prefix + "bit_xor");
+  registerBitwiseXorAggregate(prefix + "bit_xor");
 }
-} // namespace facebook::velox::functions::sparksql::aggregates
+} // namespace facebook::velox::functions::aggregate::sparksql
